Refactor visitor list code ready for re-use in summary
